+++ src/layer/domain/router/module/fetch/xhr.test.ts
@@ -15,7 +15,6 @@ describe('Unit: layer/domain/router/module/fetch/xhr', () => {
         null,
         0,
         p => p,
-        () => '',
         new Cancellation<Error>())
         .then(m => m.fmap(res => {
           assert(res.url.href === standardize('', window.location.href));
@@ -34,7 +33,6 @@ describe('Unit: layer/domain/router/module/fetch/xhr', () => {
         null,
         1,
         p => p,
-        () => '',
         new Cancellation<Error>())
         .then(m => m.extract(err => {
           assert(err instanceof Error);
@@ -52,7 +50,6 @@ describe('Unit: layer/domain/router/module/fetch/xhr', () => {
         null,
         0,
         p => p,
-        () => '',
         new Cancellation<Error>())
         .then(m => m.fmap(() => {
           assert(Date.now() - time > 1000 - 10);
@@ -70,7 +67,6 @@ describe('Unit: layer/domain/router/module/fetch/xhr', () => {
         null,
         0,
         p => p,
-        () => '',
         cancellation)
         .then(m => m.extract(err => {
           assert(err instanceof Error);